fixed tests, generalized lucide react mock
