+++ __tests__/__mocks__/lucide-react.js
@@ -1,9 +1,11 @@
-module.exports = {
-  // mock lucide-react icons to prevent Jest from failing
-  ArrowUpRight: () => null,
-  NotebookPen: () => null,
-  SquarePen: () => null,
-  ChevronDown: () => null,
-};
+// @ts-nocheck
+// mock all lucide-react icons with empty components so don't have to create specific oness
+
+module.exports = new Proxy(
+  {},
+  {
+    get: (target, prop) => () => null,
+  },
+);
 
 test("empty test just to pass", () => {});
