+++ package.json
@@ -55,6 +55,7 @@
     "prettier": "^3",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "whatwg-fetch": "^3.6.20"
   }
 }
